+++ server/libs/modules/components/merge-helper/build.gradle.kts
