+++ server/libs/modules/components/merge-helper/src/main/java/com/bytechef/component/mergehelper/constant/MergeHelperConstants.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.component.mergehelper.constant;
+
+public class MergeHelperConstants {
+
+    public static final String COMBINE_BY = "combineBy";
+    public static final String FIELD_TO_MATCH = "fieldToMatch";
+    public static final String INPUTS = "inputs";
+    public static final String INPUT_1 = "input1";
+    public static final String INPUT_2 = "input2";
+    public static final String SQL_QUERY = "sqlQuery";
+
+    private MergeHelperConstants() {
+    }
+}
