+++ server/libs/modules/components/merge-helper/src/test/java/com/bytechef/component/mergehelper/MergeHelperComponentHandlerTest.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.component.mergehelper;
+
+import com.bytechef.test.jsonasssert.JsonFileAssert;
+import org.junit.jupiter.api.Test;
+
+class MergeHelperComponentHandlerTest {
+
+    @Test
+    void testGetComponentDefinition() {
+        JsonFileAssert.assertEquals("definition/mergeHelper_v1.json",
+            new MergeHelperComponentHandler().getDefinition());
+    }
+}
