+++ settings.gradle.kts
@@ -375,7 +375,7 @@ include("server:libs:modules:components:map")
 include("server:libs:modules:components:math-helper")
 include("server:libs:modules:components:mattermost")
 include("server:libs:modules:components:mautic")
-include("server:libs:modules:components:merge")
+include("server:libs:modules:components:merge-helper")
 include("server:libs:modules:components:mcp-client")
 include("server:libs:modules:components:microsoft:microsoft-commons")
 include("server:libs:modules:components:microsoft:microsoft-dynamics-crm")
